Add mention data conversion (cont.)
