+++ packages/ckeditor5-markdown-gfm/src/gfmdataprocessor.js
@@ -8,6 +8,7 @@
  */
 
 import { HtmlDataProcessor } from 'ckeditor5/src/engine';
+import escape from 'lodash-es/escape';
 
 import markdown2html from './markdown2html/markdown2html';
 import html2markdown, { turndownService } from './html2markdown/html2markdown';
@@ -25,7 +26,7 @@ export default class GFMDataProcessor {
 	 *
 	 * @param {module:engine/view/document~Document} document
 	 */
-	constructor( document ) {
+	constructor( document, { mentionIdToText } ) {
 		/**
 		 * HTML data processor used to process HTML produced by the Markdown-to-HTML converter and the other way.
 		 *
@@ -33,6 +34,8 @@ export default class GFMDataProcessor {
 		 * @member {module:engine/dataprocessor/htmldataprocessor~HtmlDataProcessor}
 		 */
 		this._htmlDP = new HtmlDataProcessor( document );
+
+		this.mentionIdToText = mentionIdToText;
 	}
 
 	/**
@@ -55,7 +58,16 @@ export default class GFMDataProcessor {
 	 */
 	toView( data ) {
 		const html = markdown2html( data );
-		return this._htmlDP.toView( html );
+
+		// Converts `@:5f7e99c932cb8c00061b87d9:` to `<span class="mention" data-mention="@5f7e99c932cb8c00061b87d9">@Patrick Star</span>`
+		const htmlWithMentions = html.replace(
+			/@:([0-9a-fA-F]{24}):/g,
+			( match, id ) => `<span class="mention" data-mention="@${ id }">@${
+				escape( typeof this.mentionIdToText === 'function' ? this.mentionIdToText( id ) : id )
+			}</span>`
+		);
+
+		return this._htmlDP.toView( htmlWithMentions );
 	}
 
 	/**
@@ -67,7 +79,32 @@ export default class GFMDataProcessor {
 	 */
 	toData( viewFragment ) {
 		const html = this._htmlDP.toData( viewFragment );
-		return html2markdown( html );
+
+		// Converts `<span class="mention" data-mention="@5f7e99c932cb8c00061b87d9">@Patrick Star</span>` to `@:5f7e99c932cb8c00061b87d9:`
+		// See https://github.com/taskworld/ckeditor5/blob/84700cd93725130eb28abc1657b058a62b1d662d/packages/ckeditor5-mention/src/mentionediting.js#L140-L141
+		const parser = new window.DOMParser();
+		const holder = parser.parseFromString( html, 'text/html' );
+		const walker = holder.createTreeWalker(
+			holder.body,
+			window.NodeFilter.SHOW_ELEMENT,
+			{
+				acceptNode: node =>
+					( node.classList.contains( 'mention' ) && node.getAttribute( 'data-mention' ) ) ?
+						window.NodeFilter.FILTER_ACCEPT : window.NodeFilter.FILTER_SKIP
+			}
+		);
+		let node;
+		while ( true ) {
+			node = walker.nextNode();
+			if ( node === null ) {
+				break;
+			}
+
+			// See https://github.com/taskworld/tw-backend/blob/2c30299fe57fd41a43ca4990377b8503a23266c8/app/legacy/api/message-service/MentionExtractor.ts#L9
+			node.innerText = '@:' + node.getAttribute( 'data-mention' ).replace( /^@/, '' ) + ':';
+		}
+
+		return html2markdown( holder.body.innerHTML );
 	}
 
 	/**
@@ -88,5 +125,5 @@ export default class GFMDataProcessor {
 	 * This method does not have any effect on the data processor result. It exists for compatibility with the
 	 * {@link module:engine/dataprocessor/dataprocessor~DataProcessor `DataProcessor` interface}.
 	 */
-	useFillerType() {}
+	useFillerType() { }
 }
